Resolving several warnings that come up when using stricter compiler flags (#321)

+++ Libraries/D3D12RaytracingFallback/src/CpuBVH2Builder.cpp
@@ -401,7 +401,7 @@ namespace FallbackLayer
         std::deque<StackItem*>   fifoRights;
 
         StackItem* temp = new StackItem;
-        temp->parentIndex = -1;
+        temp->parentIndex = (UINT)-1;
         temp->primitiveMetaData = primitiveMetaData;
         temp->right = false;
         temp->axis = 0;
